paramLib.c: more of paramIndex is unsigned
The check for paramIndex < 0 gives a warning on some compilers,
as paramIndex is an unsigned integer and can never be < 0.
This time the check is needed after a substraction.
cast the index into a "long" before checking for < 0.

+++ motorApp/MotorSrc/paramLib.c
@@ -136,7 +136,7 @@ static int paramSetDouble( PARAMS params, paramIndex index, double value )
     int status = PARAM_ERROR;
 
     index -= params->startVal;
-    if (index >=0 && index < params->nvals)
+    if ((long)index >=0 && index < params->nvals)
     {
         if ( params->vals[index].type != paramDouble ||
              params->vals[index].data.dval != value )
